+++ RussianSpotify.API.Core/Services/Filters/AuthorFilters/AuthorNameFilter.cs
@@ -13,6 +13,6 @@ public class AuthorNameFilter : IFilter<User>
     public Task<IOrderedQueryable<User>> FilterAsync(IQueryable<User> queryable, string filterValue,
         CancellationToken cancellationToken)
         => Task.FromResult(queryable
-            .Where(author => author.UserName!.ToLower().Contains(filterValue.ToLower()))
+            .Where(author => author.UserName.ToLower().Contains(filterValue.ToLower()))
             .OrderBy(i => i.UserName));
 }
