+++ RussianSpotify.API.Data.PostgreSQL/Seeder/DbSeeder.cs
@@ -27,7 +27,7 @@ public class DbSeeder : IDbSeeder
             Roles.UserRoleName
     };
 
-    private static readonly List<CategoryType> _baseCategories = new()
+    private static readonly List<CategoryType> BaseCategories = new()
     {
         CategoryType.HipHop,
         CategoryType.Metall,
@@ -49,7 +49,7 @@ public class DbSeeder : IDbSeeder
         var existsRolesInDb = await dbContext.Roles
             .Where(x => BaseRoles.Keys.Contains(x.Id))
             .Select(x => x.Id)
-            .ToListAsync(cancellationToken) ?? new List<Guid>();
+            .ToListAsync(cancellationToken);
 
         var rolesToSeed = BaseRoles
             .Where(x => !existsRolesInDb.Contains(x.Key))
@@ -93,13 +93,13 @@ public class DbSeeder : IDbSeeder
             .ToListAsync(cancellationToken);
 
         var categoriesToDelete = existsCategoriesInDb
-            .Where(categoryFromDb => _baseCategories.All(x => x != categoryFromDb.CategoryName))
+            .Where(categoryFromDb => BaseCategories.All(x => x != categoryFromDb.CategoryName))
             .ToList();
 
         foreach (var categoryToDelete in categoriesToDelete)
             dbContext.Categories.Remove(categoryToDelete);
 
-        _baseCategories.ForEach(x =>
+        BaseCategories.ForEach(x =>
         {
             if (existsCategoriesInDb.All(y => y.CategoryName != x))
             {
