+++ RussianSpotify.API.Shared/Interfaces/IFileControllerHelper.cs
@@ -2,7 +2,6 @@
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using RussianSpotify.API.Files.Requests.File;
 using RussianSpotify.API.Shared.Requests.File;
 
 #endregion
