+++ RussianSpotify.API.Shared/Services/FileControllerHelper.cs
@@ -3,7 +3,6 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Net.Http.Headers;
-using RussianSpotify.API.Files.Requests.File;
 using RussianSpotify.API.Shared.Interfaces;
 using RussianSpotify.API.Shared.Requests.File;
